Move into separate test
